Unify configuration and move configuration handling
