+++ src/lib/config/log.rs
@@ -1,24 +1,5 @@
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
-use std::net::Ipv4Addr;
-
-#[derive(Serialize, Deserialize)]
-#[serde(default)]
-pub struct Config {
-    pub host: Ipv4Addr,
-    pub port: u16,
-    pub log: LogConfig,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            host: Ipv4Addr::new(127, 0, 0, 1),
-            port: 8080,
-            log: LogConfig::default(),
-        }
-    }
-}
 
 #[derive(Serialize, Deserialize)]
 #[serde(default)]
