+++ src/lib/util/config.rs
@@ -1,23 +1,7 @@
-use crate::util::shell::Shell;
-use crate::web::config::Config as WebConfig;
-use clap::ArgMatches;
-use serde::{Deserialize, Serialize};
+use crate::config::Config;
+use crate::util::command::*;
 use std::error::Error;
 use std::net::Ipv4Addr;
-
-#[derive(Serialize, Deserialize)]
-#[serde(default)]
-pub struct Config {
-    pub web: WebConfig,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            web: WebConfig::default(),
-        }
-    }
-}
 
 pub fn load(path: &str) -> Result<Config, Box<dyn Error + 'static>> {
     let string = std::fs::read_to_string(path)?;
