+++ src/lib/util/mod.rs
@@ -1,2 +1,3 @@
 pub mod command;
+pub mod config;
 pub mod shell;
